+++ src/index.ejs
@@ -12,7 +12,7 @@
 <body>
 	<noscript>Please enable JavaScript in your web browser to view this page.</noscript>
 	<div id="pixi-content" class="game-container">
-        <canvas id="pixi-canvas" class="game-canvas" />
+        <canvas id="pixi-canvas" class="game-canvas">
     </div>
 </body>
